+++ addon/components/object-list-view.js
@@ -120,6 +120,14 @@ export default FlexberryBaseComponent.extend(
   sortByColumn: 'sortByColumn',
 
   /**
+    Default left padding in cells.
+    @property defaultLeftPadding
+    @type Number
+    @default 10
+  */
+  defaultLeftPadding: 10,
+
+  /**
     Flag indicates whether to look for changes of model (and displaying corresponding changes on control) or not.
 
     If flag is enabled component compares current detail array with used on component,