+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+* `flexberry-objectlistview` component:
+    * Add hierarchy support on `edit-form`.
+
+### Changed
+* `flexberry-objectlistview` component:
+    * Change hierarchy indent setting (`hierarchicalIndent` property) to `Number` type (in pixels). Also added `defaultLeftPadding` property as padding value in each level of hierarchy (in pixels).
+
+
 ## [0.8.0-beta.27] - 2016-12-29
 ### Fixed
 * Default ordering for `flexberry-lookup` component is not applying after changing current list page.